+++ src/backend/Kys.Application/Environments/Commands/CreateCustomerEnvironment/CreateCustomerEnvironmentCommandHandler.cs
@@ -25,6 +25,7 @@ public sealed class CreateCustomerEnvironmentCommandHandler(
         {
             CustomerProductId = customerProduct.Id,
             EnvironmentTypeId = envType.Id,
+            HostingPlatformId = request.HostingPlatformId,
             Name = string.IsNullOrWhiteSpace(request.Name) ? envType.Name : request.Name,
             Notes = request.Notes
         };